Finally got saving restaurants to work
First added attribute names to the html to be accessible in the javascript. Then, once the save button is clicked, these attributes are obtained. Then, the saveRestaurant function in the JS is called, which makes an API call to the database. CSRFtoken is used for validation, and that is gotten from the cookies on the website.  Once the API call is successful, it is added to a SQL lite database that is built into Django!