Added a simple page to display a users saved restaurants
I created a simple page on the index to display a user's saved restaurants. I added it to the admin.py file so the sql lite database is visible from the admin page. I added a function to the authentication views called saved_restaurants. That checks the model for values and displays them.