added functionality for delete button
Added functionality for delete button in saved_restaurants. Makes a post call to the frontend fetches the yelp_id; with the yelp_id, we then query the database and delete it if it exits. I had a problem for a while because I didn't know I had to unpack the yelp_id as a JSON object.